requirements.txt
fastapi
uvicorn
aider-chat
pytest